src/research_generation/research_gap_engine.py
import numpy as np
import networkx as nx
import pandas as pd
from typing import Dict, List, Any, Optional, Tuple
from dataclasses import dataclass, field

from src.knowledge_graph.distributed_knowledge_graph import DistributedKnowledgeGraphEngine
from src.semantic_understanding.semantic_intent_engine import SemanticIntentEngine

@dataclass
class ResearchProposal:
    """
    Represents a structured research proposal
    """
    id: str
    title: str
    domain: str
    priority: float
    novelty_score: float
    interdisciplinary_potential: float
    research_questions: List[str] = field(default_factory=list)
    proposed_methodologies: List[str] = field(default_factory=list)
    potential_impact: Dict[str, float] = field(default_factory=dict)
    confidence_interval: Tuple[float, float] = (0.0, 0.0)

class ResearchGapEngine:
    def __init__(
        self, 
        knowledge_graph: Optional[DistributedKnowledgeGraphEngine] = None,
        semantic_engine: Optional[SemanticIntentEngine] = None
    ):
        """
        Initialize Research Gap Analysis Engine
        
        Args:
            knowledge_graph: Distributed knowledge graph engine
            semantic_engine: Semantic intent engine
        """
        # Initialize dependencies
        self.knowledge_graph = knowledge_graph or DistributedKnowledgeGraphEngine()
        self.semantic_engine = semantic_engine or SemanticIntentEngine()
        
        # Research domain configuration
        self.research_domains = [
            'technology', 'science', 'engineering', 
            'mathematics', 'computing', 'interdisciplinary'
        ]
        
        # Research priority scoring parameters
        self.priority_weights = {
            'novelty': 0.4,
            'impact_potential': 0.3,
            'interdisciplinary_potential': 0.2,
            'research_complexity': 0.1
        }
        
        # Gap detection configuration
        self.gap_detection_threshold = 0.7
    
    def identify_research_gaps(
        self, 
        domain: Optional[str] = None,
        max_proposals: int = 5
    ) -> List[ResearchProposal]:
        """
        Identify potential research gaps across domains
        
        Args:
            domain: Specific domain to analyze (optional)
            max_proposals: Maximum number of research proposals to generate
        
        Returns:
            List of research proposals representing potential research gaps
        """
        # Use domain or analyze all domains
        domains_to_analyze = [domain] if domain else self.research_domains
        
        research_proposals = []
        
        for current_domain in domains_to_analyze:
            # Analyze knowledge graph for research gaps
            domain_gaps = self._analyze_domain_gaps(current_domain)
            
            # Generate research proposals for detected gaps
            domain_proposals = self._generate_research_proposals(
                current_domain, 
                domain_gaps
            )
            
            research_proposals.extend(domain_proposals)
        
        # Sort and select top proposals
        research_proposals.sort(key=lambda x: x.priority, reverse=True)
        
        return research_proposals[:max_proposals]
    
    def _analyze_domain_gaps(self, domain: str) -> List[Dict[str, Any]]:
        """
        Analyze research gaps within a specific domain
        
        Args:
            domain: Domain to analyze
        
        Returns:
            List of detected research gaps
        """
        # Extract domain-specific knowledge graph
        domain_graph = self.knowledge_graph.extract_domain_subgraph(domain)
        
        # Compute node centrality to identify potential research gaps
        centrality_scores = nx.betweenness_centrality(domain_graph)
        
        # Identify low-centrality nodes as potential research gaps
        research_gaps = [
            {
                'node': node,
                'centrality': score,
                'metadata': domain_graph.nodes[node]
            }
            for node, score in centrality_scores.items()
            if score < self.gap_detection_threshold
        ]
        
        return research_gaps
    
    def _generate_research_proposals(
        self, 
        domain: str, 
        research_gaps: List[Dict[str, Any]]
    ) -> List[ResearchProposal]:
        """
        Generate research proposals for detected gaps
        
        Args:
            domain: Research domain
            research_gaps: List of detected research gaps
        
        Returns:
            Generated research proposals
        """
        proposals = []
        
        for gap in research_gaps:
            # Generate unique proposal ID
            proposal_id = f"proposal_{hash(str(gap))}"
            
            # Compute novelty and interdisciplinary potential
            novelty_score = self._compute_novelty_score(gap)
            interdisciplinary_potential = self._assess_interdisciplinary_potential(gap)
            
            # Generate research questions
            research_questions = self._generate_research_questions(gap)
            
            # Generate proposed methodologies
            proposed_methodologies = self._suggest_research_methodologies(gap)
            
            # Compute priority and potential impact
            priority, potential_impact = self._compute_research_priority(
                novelty_score, 
                interdisciplinary_potential
            )
            
            # Create research proposal
            proposal = ResearchProposal(
                id=proposal_id,
                title=f"Exploring {gap['node']} in {domain} Research",
                domain=domain,
                priority=priority,
                novelty_score=novelty_score,
                interdisciplinary_potential=interdisciplinary_potential,
                research_questions=research_questions,
                proposed_methodologies=proposed_methodologies,
                potential_impact=potential_impact,
                confidence_interval=(priority - 0.2, priority + 0.2)
            )
            
            proposals.append(proposal)
        
        return proposals
    
    def _compute_novelty_score(self, gap: Dict[str, Any]) -> float:
        """
        Compute novelty score for a research gap
        
        Args:
            gap: Research gap information
        
        Returns:
            Novelty score
        """
        # Compute novelty based on graph connectivity and metadata
        node_metadata = gap.get('metadata', {})
        connectivity = gap.get('centrality', 0.0)
        
        # Novelty factors
        metadata_novelty = len(node_metadata) / 10.0  # Normalize
        connectivity_factor = 1 - connectivity
        
        # Combine novelty factors
        novelty_score = 0.6 * connectivity_factor + 0.4 * metadata_novelty
        
        return max(0.0, min(1.0, novelty_score))
    
    def _assess_interdisciplinary_potential(self, gap: Dict[str, Any]) -> float:
        """
        Assess interdisciplinary potential of a research gap
        
        Args:
            gap: Research gap information
        
        Returns:
            Interdisciplinary potential score
        """
        # Analyze node metadata for cross-domain connections
        node_metadata = gap.get('metadata', {})
        
        # Check for domain-related keywords
        interdisciplinary_keywords = [
            'cross-domain', 'interdisciplinary', 'multi-modal', 
            'hybrid', 'integrated', 'collaborative'
        ]
        
        # Compute interdisciplinary score
        keyword_match = sum(
            1 for keyword in interdisciplinary_keywords 
            if any(keyword in str(value).lower() for value in node_metadata.values())
        )
        
        interdisciplinary_score = keyword_match / len(interdisciplinary_keywords)
        
        return max(0.0, min(1.0, interdisciplinary_score))
    
    def _generate_research_questions(self, gap: Dict[str, Any]) -> List[str]:
        """
        Generate potential research questions for a gap
        
        Args:
            gap: Research gap information
        
        Returns:
            List of research questions
        """
        node = gap.get('node', 'Unknown')
        
        # Generate research questions using semantic patterns
        research_question_templates = [
            f"How can we advance understanding of {node}?",
            f"What are the underlying mechanisms of {node}?",
            f"What innovative approaches can be developed for {node}?",
            f"What are the potential limitations and challenges in {node}?"
        ]
        
        return research_question_templates
    
    def _suggest_research_methodologies(self, gap: Dict[str, Any]) -> List[str]:
        """
        Suggest research methodologies for a gap
        
        Args:
            gap: Research gap information
        
        Returns:
            List of proposed research methodologies
        """
        # Methodology suggestions based on domain and gap characteristics
        methodology_suggestions = [
            "Systematic literature review",
            "Empirical data collection and analysis",
            "Computational modeling",
            "Experimental prototype development",
            "Comparative case study approach"
        ]
        
        return methodology_suggestions
    
    def _compute_research_priority(
        self, 
        novelty_score: float, 
        interdisciplinary_potential: float
    ) -> Tuple[float, Dict[str, float]]:
        """
        Compute research priority and potential impact
        
        Args:
            novelty_score: Novelty of the research gap
            interdisciplinary_potential: Interdisciplinary potential
        
        Returns:
            Research priority and potential impact
        """
        # Compute priority using weighted factors
        priority = (
            self.priority_weights['novelty'] * novelty_score +
            self.priority_weights['interdisciplinary_potential'] * interdisciplinary_potential +
            self.priority_weights['research_complexity'] * np.random.random() +
            self.priority_weights['impact_potential'] * np.random.random()
        )
        
        # Potential impact estimation
        potential_impact = {
            'scientific_advancement': priority * 0.4,
            'technological_innovation': priority * 0.3,
            'economic_potential': priority * 0.2,
            'societal_impact': priority * 0.1
        }
        
        return max(0.0, min(1.0, priority)), potential_impact
    
    def generate_research_proposal(
        self, 
        proposal: ResearchProposal
    ) -> Dict[str, Any]:
        """
        Generate a comprehensive research proposal document
        
        Args:
            proposal: Research proposal to elaborate
        
        Returns:
            Detailed research proposal document
        """
        # Elaborate on research proposal
        proposal_document = {
            'id': proposal.id,
            'title': proposal.title,
            'abstract': self._generate_proposal_abstract(proposal),
            'research_context': self._generate_research_context(proposal),
            'methodology': {
                'research_questions': proposal.research_questions,
                'proposed_methodologies': proposal.proposed_methodologies
            },
            'expected_outcomes': self._generate_expected_outcomes(proposal),
            'impact_assessment': proposal.potential_impact,
            'confidence_interval': proposal.confidence_interval
        }
        
        return proposal_document
    
    def _generate_proposal_abstract(self, proposal: ResearchProposal) -> str:
        """
        Generate an abstract for the research proposal
        
        Args:
            proposal: Research proposal
        
        Returns:
            Proposal abstract
        """
        return (
            f"This research proposal explores critical gaps in {proposal.domain} "
            f"with a focus on novel approaches to {proposal.title}. "
            f"With a novelty score of {proposal.novelty_score:.2f} and "
            f"interdisciplinary potential of {proposal.interdisciplinary_potential:.2f}, "
            "this study aims to push the boundaries of current understanding."
        )
    
    def _generate_research_context(self, proposal: ResearchProposal) -> Dict[str, Any]:
        """
        Generate research context for the proposal
        
        Args:
            proposal: Research proposal
        
        Returns:
            Research context details
        """
        return {
            'domain': proposal.domain,
            'research_landscape': f"Emerging challenges in {proposal.domain}",
            'current_limitations': "Existing knowledge gaps and research constraints"
        }
    
    def _generate_expected_outcomes(self, proposal: ResearchProposal) -> Dict[str, Any]:
        """
        Generate expected research outcomes
        
        Args:
            proposal: Research proposal
        
        Returns:
            Expected research outcomes
        """
        return {
            'scientific_contributions': [
                f"Advance understanding of {proposal.title}",
                "Develop novel methodological approaches"
            ],
            'practical_implications': [
                "Potential technological innovations",
                "Insights for future research directions"
            ],
            'knowledge_transfer_potential': proposal.potential_impact
        }
